apps/memory-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
mem0ai>=1.0.0
pydantic==2.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.10
google-genai>=1.0.0
